+++ packages/hardhat-core/test/fixture-projects/unlimited-contract-size/hardhat.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-  networks: {
-    hardhat: {
-      allowUnlimitedContractSize: true,
-    },
-  },
-  solidity: "0.5.15",
-};