+++ packages/hardhat-ledger/src/provider.ts
@@ -35,6 +35,7 @@ export class LedgerProvider extends ProviderWrapperWithChainId {
 
   public readonly paths: Paths = {}; // { address: path }
   public name: string = "LedgerProvider";
+  public isOutputEnabled: boolean = true;
 
   protected _eth: EthWrapper | undefined;
 
@@ -113,6 +114,10 @@ export class LedgerProvider extends ProviderWrapperWithChainId {
 
   public async request(args: RequestArguments): Promise<unknown> {
     const params = this._getParams(args);
+
+    if (args.method === "hardhat_setLedgerOutputEnabled") {
+      return this._setOutputEnabled(params);
+    }
 
     if (args.method === "eth_accounts") {
       const accounts = (await this._wrappedProvider.request(args)) as string[];
@@ -341,7 +346,7 @@ export class LedgerProvider extends ProviderWrapperWithChainId {
         account <= LedgerProvider.MAX_DERIVATION_ACCOUNTS;
         account++
       ) {
-        path = `44'/60'/${account}'/0/0`;
+        path = this._getDerivationPath(account);
 
         this.emit("derivation_progress", path, account);
 
@@ -369,9 +374,21 @@ export class LedgerProvider extends ProviderWrapperWithChainId {
 
     this.emit("derivation_failure");
     throw new HardhatLedgerDerivationPathError(
-      `Could not find a valid derivation path for ${addressToFind}. Paths from m/44'/60'/0'/0/0 to m/44'/60'/${LedgerProvider.MAX_DERIVATION_ACCOUNTS}'/0/0 were searched.`,
+      `Could not find a valid derivation path for ${addressToFind}. Paths from ${this._getDerivationPath(
+        0
+      )} to ${this._getDerivationPath(
+        LedgerProvider.MAX_DERIVATION_ACCOUNTS
+      )} were searched.`,
       path
     );
+  }
+
+  private _getDerivationPath(index: number): string {
+    if (this.options.derivationFunction === undefined) {
+      return `m/44'/60'/${index}'/0/0`;
+    } else {
+      return this.options.derivationFunction(index);
+    }
   }
 
   private async _withConfirmation<T extends (...args: any) => any>(
@@ -427,5 +444,15 @@ export class LedgerProvider extends ProviderWrapperWithChainId {
         hexAddress
       );
     }
+  }
+
+  /**
+   * Toggles the provider's output. Use to suppress default feedback and
+   * manage it via events.
+   */
+  private _setOutputEnabled(params: any[]): void {
+    const [enabled] = validateParams(params, t.boolean);
+
+    this.isOutputEnabled = enabled;
   }
 }