+++ packages/hardhat-vyper/test/fixture-projects/compilation-with-settings-option-variants/optimize-true-not-available-old-versions/hardhat.config.js
@@ -1,0 +1,15 @@
+require("../../../../src/index");
+
+module.exports = {
+  vyper: {
+    compilers: [
+      {
+        version: "0.3.0",
+        settings: {
+          evmVersion: "paris",
+          optimize: true,
+        },
+      },
+    ],
+  },
+};