+++ v-next/hardhat-build-system/src/internal/solidity/compiler/downloader.ts
@@ -9,6 +9,7 @@ import {
 } from "@nomicfoundation/hardhat-errors";
 import { bytesToHexString } from "@nomicfoundation/hardhat-utils/bytes";
 import { keccak256 } from "@nomicfoundation/hardhat-utils/crypto";
+import { ensureError } from "@nomicfoundation/hardhat-utils/error";
 import {
   chmod,
   createFile,
@@ -197,7 +198,9 @@ export class CompilerDownloader implements ICompilerDownloader {
       if (build === undefined && (await this.#shouldDownloadCompilerList())) {
         try {
           await this.#downloadCompilerList();
-        } catch (e: any) {
+        } catch (e) {
+          ensureError(e);
+
           throw new HardhatError(
             ERRORS.SOLC.VERSION_LIST_DOWNLOAD_FAILED,
             {},
@@ -217,7 +220,9 @@ export class CompilerDownloader implements ICompilerDownloader {
       let downloadPath: string;
       try {
         downloadPath = await this.#downloadCompiler(build);
-      } catch (e: any) {
+      } catch (e) {
+        ensureError(e);
+
         throw new HardhatError(
           ERRORS.SOLC.DOWNLOAD_FAILED,
           {