+++ v-next/hardhat-build-system/src/internal/utils/fs-utils.ts
@@ -3,7 +3,7 @@ import fsPromises from "node:fs/promises";
 import path from "node:path";
 
 import { assertHardhatInvariant } from "@nomicfoundation/hardhat-errors";
-import { CustomError } from "@nomicfoundation/hardhat-utils/error";
+import { ensureError, CustomError } from "@nomicfoundation/hardhat-utils/error";
 
 /**
  * Returns an array of files (not dirs) that match a condition.
@@ -41,7 +41,12 @@ export async function getAllFilesMatching(
 async function readdir(absolutePathToDir: string) {
   try {
     return await fsPromises.readdir(absolutePathToDir);
-  } catch (e: any) {
+  } catch (e) {
+    ensureError(e);
+    if (!("code" in e) || typeof e.code !== "string") {
+      throw e;
+    }
+
     if (e.code === "ENOENT") {
       return [];
     }
@@ -140,7 +145,12 @@ export function getAllFilesMatchingSync(
 function readdirSync(absolutePathToDir: string) {
   try {
     return fs.readdirSync(absolutePathToDir);
-  } catch (e: any) {
+  } catch (e) {
+    ensureError(e);
+    if (!("code" in e) || typeof e.code !== "string") {
+      throw e;
+    }
+
     if (e.code === "ENOENT") {
       return [];
     }
@@ -205,7 +215,12 @@ export async function getRealPath(absolutePath: string): Promise<string> {
     // This method returns the actual casing.
     // Please read Node.js' docs to learn more.
     return await fsPromises.realpath(path.normalize(absolutePath));
-  } catch (e: any) {
+  } catch (e) {
+    ensureError(e);
+    if (!("code" in e) || typeof e.code !== "string") {
+      throw e;
+    }
+
     if (e.code === "ENOENT") {
       // eslint-disable-next-line @nomicfoundation/hardhat-internal-rules/only-hardhat-error
       throw new FileNotFoundError(absolutePath, e);
@@ -226,7 +241,12 @@ export function getRealPathSync(absolutePath: string): string {
     // This method returns the actual casing.
     // Please read Node.js' docs to learn more.
     return fs.realpathSync.native(path.normalize(absolutePath));
-  } catch (e: any) {
+  } catch (e) {
+    ensureError(e);
+    if (!("code" in e) || typeof e.code !== "string") {
+      throw e;
+    }
+
     if (e.code === "ENOENT") {
       // eslint-disable-next-line @nomicfoundation/hardhat-internal-rules/only-hardhat-error
       throw new FileNotFoundError(absolutePath, e);