+++ v-next/hardhat-build-system/test/helpers.ts
@@ -6,6 +6,7 @@ import { beforeEach } from "node:test";
 import { fileURLToPath, pathToFileURL } from "node:url";
 
 import { ErrorDescriptor, HardhatError } from "@nomicfoundation/hardhat-errors";
+import { ensureError } from "@nomicfoundation/hardhat-utils/error";
 import { getRealPath, mkdir, remove } from "@nomicfoundation/hardhat-utils/fs";
 import semver from "semver";
 
@@ -95,7 +96,12 @@ export async function importCsjOrEsModule(filePath: string): Promise<any> {
   try {
     const imported = await import(pathToFileURL(filePath).toString());
     return imported.default !== undefined ? imported.default : imported;
-  } catch (e: any) {
+  } catch (e) {
+    ensureError(e);
+    if (!("code" in e) || typeof e.code !== "string") {
+      throw e;
+    }
+
     if (e.code === "ERR_REQUIRE_ESM") {
       throw new Error("Cannot find configuration file");
     }
@@ -212,7 +218,7 @@ export async function expectHardhatErrorAsync(
 
   try {
     await f();
-  } catch (err: unknown) {
+  } catch (err) {
     if (!HardhatError.isHardhatError(err)) {
       assert.fail();
     }